+++ ui/__init__.py
@@ -1,1 +1,0 @@
-"""Ui package for MRT Pharma Digital Twin."""