+++ ui/layout.py
@@ -1,0 +1,2 @@
+def configure_page(st):
+    st.set_page_config(page_title='MRT Pharma Digital Twin',layout='wide')